backend/requirements.txt
Django==4.2.9
psycopg[binary]
tzdata
djangorestframework
djangorestframework-simplejwt
drf-spectacular
